% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset_singleRep}
\alias{dataset_singleRep}
\title{\code{dataset_singleRep} is a dataset (matrix) containing synthetic, example data to demostrate the format that a data matrix should take when only a single large cage experiment was used.}
\description{
\code{dataset_singleRep} is a dataset (matrix) containing synthetic, example data to demostrate the format that a data matrix should take when only a single large cage experiment was used.
}
\details{
Each row  of the matrix contains data for each female recovered from a mating cage.  The three columns are: (i) "Cage" which contains an integer indicating which replicate cage the female came from; (ii) "NumEggs" which is the number of eggs laid by each female; and (iii) "NumViable" which is the number of eggs that hatched or produced viable offspring.
}
\author{
Dan Pagendam \email{dan.pagendam@csiro.au}
}
\keyword{data}
