% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fade.R
\name{fade}
\alias{fade}
\title{A function for creating transparent colours}
\usage{
fade(colors, alpha)
}
\arguments{
\item{\code{colors}}{is a character string of colour names to which transparency should be added.}

\item{\code{alpha}}{is a numeric vector specifying the desired opactity for each colour in \code{colors}.  This should be an integer between 0 and 255}
}
\value{
The function returns a string containing the hexidecimal colour code and is then easily used as the colour to use in standard R plot commands.
}
\description{
\code{fade} provides the name of a semi-transparent R colour after passing the name of a base colour or colours (e.g. "blue") and the degree of opacity desired.
}
\examples{
newColour <- fade(c("blue", "green", "purple"), c(50, 100, 200))
}
